switch to rewritten asyncAtom
