+++ src/utils/atoms/createAsyncAtom/createAsyncAtom.old.test.ts
@@ -2,7 +2,7 @@ import { createAtom, createStore } from '@reatom/core-v2';
 import { expect, test, describe, vi, beforeEach } from 'vitest';
 import { createBooleanAtom } from '@reatom/core-v2/primitives';
 import { incrementId, wait, waitMockCalls } from '~utils/test';
-import { createAsyncAtom } from './createAsyncAtom.next';
+import { createAsyncAtom } from './createAsyncAtom.old';
 import { ABORT_ERROR_MESSAGE } from './abort-error';
 import type { Store } from '@reatom/core-v2';
 
@@ -34,7 +34,7 @@ describe('Resource atom add resource state structure', () => {
     });
   });
 
-  test('have correct loading state', async ({ store }) => {
+  test('have correct loading state', ({ store }) => {
     const resAtomA = createAsyncAtom(null, async () => await wait(1), id(), {
       store,
     });
@@ -46,16 +46,10 @@ describe('Resource atom add resource state structure', () => {
   });
 
   test('have correct lastParams state', async ({ store }) => {
-    const resAtomA = createAsyncAtom(
-      null,
-      async (_, abortController) => await wait(1),
-      id(),
-      {
-        store,
-      },
-    );
+    const resAtomA = createAsyncAtom(null, async () => await wait(1), id(), {
+      store,
+    });
     store.dispatch(resAtomA.request('foo'));
-    await wait(0.1);
     expect(resAtomA.getState()).toMatchObject({
       lastParams: 'foo',
     });
@@ -68,10 +62,10 @@ describe('Resource atom add resource state structure', () => {
         /* Throw error when 'bad' in params */
         await wait(1, params === 'bad' ? { failWithMessage: 'Test error' } : {}),
       'resAtomA',
-      { store, auto: false },
+      { store },
     );
     store.dispatch(resAtomA.request('bad'));
-    await wait(2);
+    await wait(1);
     expect(resAtomA.getState().error).toBe('Test error');
     store.dispatch(resAtomA.request('good'));
     await wait(1);
@@ -86,7 +80,7 @@ describe('Resource atom add resource state structure', () => {
         return 1;
       },
       id(),
-      { store, auto: false },
+      { store },
     );
     store.dispatch(resAtomA.request('foo'));
     await wait(1);
@@ -111,7 +105,6 @@ describe('Resource canceling', () => {
       id(),
       {
         store,
-        auto: false,
       },
     );
     store.dispatch(resAtomA.request(1));
@@ -241,40 +234,6 @@ describe('Resource canceling', () => {
     });
   });
 
-  test('Resource not wait aborted request for continue work', async ({ store }) => {
-    const stateChangesLog = vi.fn(async (arg) => null);
-
-    const resAtomA = createAsyncAtom(
-      null,
-      async (value) => {
-        // timeout longer than available
-        await wait(5);
-        return value;
-      },
-      id(),
-      {
-        store,
-        auto: false,
-      },
-    );
-
-    resAtomA.subscribe((s) => stateChangesLog(s));
-
-    store.dispatch(resAtomA.request(1));
-    await wait(1);
-    store.dispatch(resAtomA.request(2));
-
-    await waitMockCalls(stateChangesLog, 3);
-
-    expect(stateChangesLog).toHaveBeenNthCalledWith(3, {
-      error: ABORT_ERROR_MESSAGE,
-      dirty: true,
-      data: null,
-      lastParams: 1, // should have parameters of request that was canceled
-      loading: false, // should out from loading state
-    });
-  });
-
   test('Resource set error:canceled state after canceled by other request when fetcher use try catch', async ({
     store,
   }) => {
@@ -286,7 +245,7 @@ describe('Resource canceling', () => {
         // I'am not rise any error in fetcher on cancel
         // It's the similar to wrap real fetcher in try catch
         // Because our client rise error on cancel signal automatically
-        await wait(3);
+        await wait(5);
         return value;
       },
       id(),
@@ -424,7 +383,7 @@ describe('Resource reactivity', () => {
         return 'updated-' + ++i;
       },
       id(),
-      { store, auto: false },
+      { store },
     );
 
     // listen changes
@@ -433,22 +392,23 @@ describe('Resource reactivity', () => {
 
     // mutate deps
     store.dispatch(deps.set(true));
-    expect(stateChangesLog).toHaveBeenNthCalledWith(2, {
+
+    expect(stateChangesLog).toHaveBeenNthCalledWith(1, {
       error: null,
       dirty: true,
       data: null,
-      lastParams: true,
+      lastParams: false,
       loading: true,
     });
-    await wait(1);
 
-    store.dispatch(deps.set(false));
+    await waitMockCalls(stateChangesLog, 4);
+
     expect(stateChangesLog).toHaveBeenNthCalledWith(4, {
       error: null,
       dirty: true,
-      data: 'updated-1',
-      lastParams: false,
-      loading: true,
+      data: 'updated-2',
+      lastParams: true,
+      loading: false,
     });
   });
 
@@ -564,7 +524,7 @@ describe('Resource atoms chaining state', () => {
         await wait(1);
       },
       id(),
-      { store, auto: false },
+      { store },
     );
 
     const resAtomB = createAsyncAtom(
@@ -573,12 +533,11 @@ describe('Resource atoms chaining state', () => {
         await wait(1);
       },
       id(),
-      { store, inheritState: true },
+      { store },
     );
 
     store.dispatch(resAtomA.request(null));
-    await wait(0.1);
-
+    await wait(1);
     expect(resAtomB.getState().loading).toBe(true);
   });
 
@@ -589,7 +548,7 @@ describe('Resource atoms chaining state', () => {
         await wait(1, { failWithMessage: 'Test error' });
       },
       id(),
-      { store, auto: false },
+      { store },
     );
 
     const resAtomB = createAsyncAtom(
@@ -603,7 +562,6 @@ describe('Resource atoms chaining state', () => {
 
     store.dispatch(resAtomA.request(null));
     await wait(1);
-
     expect(resAtomB.getState().error).toBe('Test error');
   });
 
@@ -615,7 +573,7 @@ describe('Resource atoms chaining state', () => {
         return 'result';
       },
       id(),
-      { store, auto: false },
+      { store },
     );
 
     const resAtomB = createAsyncAtom(
@@ -627,8 +585,7 @@ describe('Resource atoms chaining state', () => {
       { store, inheritState: false },
     );
 
-    store.dispatch(resAtomA.request(null));
-    await wait(0.1);
+    await wait(1);
 
     expect(resAtomA.getState().loading).toBe(true);
     expect(resAtomB.getState().loading).toBe(false);
