+++ src/utils/atoms/createAsyncAtom/createAsyncAtom.ts
@@ -1,32 +1,22 @@
+import { action, atom } from '@reatom/core';
 import { isObject } from '@reatom/core-v2';
-import { createAtom } from '~utils/atoms/createPrimitives';
-import { store } from '~core/store/store';
-import { isErrorWithMessage } from '~utils/common';
-import { abortable, ABORT_ERROR_MESSAGE, isAbortError } from './abort-error';
+import { deferred, type DeferredPromise } from '@homer0/deferred';
+import { isErrorWithMessage } from '~utils/common/error';
+import { store as defaultStore } from '~core/store/store';
+import { v3toV2 } from '../v3tov2';
+import { ABORT_ERROR_MESSAGE, abortable, isAbortError } from './abort-error';
 import { isAtomLike } from './is-atom-like';
-import type { AsyncAtomOptions, AsyncAtomState, Fetcher, AsyncAtomDeps } from './types';
-import type { AtomBinded, AtomSelfBinded, AtomState } from '@reatom/core-v2';
+import type { Ctx } from '@reatom/core';
+import type { AsyncAtomDeps, AsyncAtomOptions, AsyncAtomState, Fetcher } from './types';
+import type { AtomBinded, AtomState, AtomSelfBinded } from '@reatom/core-v2';
 
-const verbose = true;
-const filterByAtomName = '';
-const logger =
-  (name: string) =>
-  (...args: Array<string>) => {
-    if (!verbose) return;
-    // eslint-disable-next-line
-    console.log(...args);
-  };
-
-type ResourceCtx = {
-  abortController?: null | AbortController;
-  activeRequest?: Promise<any>;
-};
-
-const defaultOptions: AsyncAtomOptions<never, never> = {
-  inheritState: false,
-  store: store,
-  auto: true,
-};
+function generateErrorMessage(e: unknown): string {
+  if (isAbortError(e)) {
+    return ABORT_ERROR_MESSAGE;
+  } else {
+    return isErrorWithMessage(e) ? e.message : typeof e === 'string' ? e : 'Unknown';
+  }
+}
 
 /* Check that name unique */
 const getUniqueId = ((mem) => {
@@ -43,11 +33,17 @@ const getUniqueId = ((mem) => {
   };
 })(new Set());
 
+const defaultOptions = {
+  inheritState: false,
+  store: defaultStore,
+  auto: true,
+};
+
 export function createAsyncAtom<
   F extends Fetcher<Exclude<AtomState<D>, null>, Awaited<ReturnType<F>>>,
   D extends AtomBinded,
 >(
-  atom: D | null,
+  depsAtom: D | null,
   fetcher: F,
   name: string,
   resourceAtomOptions: AsyncAtomOptions<Awaited<ReturnType<F>>, AtomState<D>> = {},
@@ -55,216 +51,168 @@ export function createAsyncAtom<
   AsyncAtomState<AtomState<D>, Awaited<ReturnType<F>>>,
   AsyncAtomDeps<D, F>
 > {
-  const log = logger(name);
+  const log = (...args) => console.debug(`[${name}]: `, ...args);
+  const logStateChange = (from, to) => {
+    console.debug(`[${name}] from -> to: `, from, to);
+    return to;
+  };
 
-  const options: AsyncAtomOptions<Awaited<ReturnType<F>>, AtomState<D>> = {
+  log('created');
+
+  type State = AsyncAtomState<AtomState<D>, Awaited<ReturnType<F>>>;
+
+  type Options = AsyncAtomOptions<Awaited<ReturnType<F>>, AtomState<D>>;
+  const options: Options & Required<Pick<Options, 'store'>> = {
     ...resourceAtomOptions,
     auto: resourceAtomOptions.auto ?? defaultOptions.auto,
     inheritState: resourceAtomOptions.inheritState ?? defaultOptions.inheritState,
     store: resourceAtomOptions.store ?? defaultOptions.store,
   };
 
-  const deps: AsyncAtomDeps<D, F> = {
-    request: (params) => params,
-    refetch: () => null,
-    cancel: () => null,
-    _done: (params, data) => ({ params, data }),
-    _error: (params, error) => ({ params, error }),
-    _loading: (params) => ({ params }),
-  };
-
-  if (atom) {
-    deps.depsAtom = atom;
-  }
-
-  const resourceAtom: AtomSelfBinded<
-    AsyncAtomState<AtomState<D>, Awaited<ReturnType<F>>>,
-    AsyncAtomDeps<D, F>
-  > = createAtom(
-    deps,
-    (
-      { onAction, schedule, create, onChange, get },
-      state: AsyncAtomState<AtomState<D>, Awaited<ReturnType<F>>> = {
-        loading: false,
-        data: null,
-        error: null,
-        lastParams: null,
-        dirty: false,
-      },
-    ) => {
-      type Context = ResourceCtx;
-      const newState = { ...state };
-
-      onAction('request', (params) => {
-        newState.dirty = true; // For unblock refetch
-        schedule(async (dispatch, ctx: Context) => {
-          log('1. Request', params);
-          // Before making new request we should abort previous request
-          // If some request active right now we have abortController
-          if (ctx.abortController) {
-            log('2. Cancel flow');
-            ctx.abortController.abort();
-            delete ctx.abortController;
-            /**
-             * In case when we cancel active request by another request
-             * without this lines we have wrong sequence
-             * 1) Loading A
-             *    (new request)
-             * 2) Loading B
-             * 3) Error - A canceled
-             *
-             * This hackish solution the only way I found to make order correct
-             * 1) Loading A
-             *    (new request)
-             * 2) Error - A canceled
-             * 3) Loading B
-             */
-            let isAfterAbort = false;
-            try {
-              log('2.1. Cancel flow: await activeRequest');
-              await ctx.activeRequest!;
-            } catch (e) {
-              log('2.2. Cancel flow: activeRequest have error');
-              isAfterAbort = isAbortError(e);
-            }
-            if (isAfterAbort) {
-              log('2.3. Cancel flow: activeRequest was canceled');
-              // This is abort transaction just for abort previous request.
-              // Move this request to next transaction
-              log('2.4. Repeat request');
-              dispatch(create('request', params));
-            } // else - just regular error, no additional actions
-            return;
-          }
-          log('3. Set loading state');
-          dispatch(create('_loading', params));
-          const abortController = new AbortController();
-          try {
-            ctx.abortController = abortController;
-            ctx.activeRequest = abortable(
-              abortController,
-              fetcher(params, abortController),
-            );
-            log('4. Wait result');
-            const fetcherResult = await ctx.activeRequest;
-            if (fetcherResult === undefined)
-              console.warn('resourceAtom: fetcherResult undefined');
-            delete ctx.activeRequest;
-            log('5. Check that it was aborted:');
-            abortController.signal.throwIfAborted(); // Alow set canceled state, even if abort error was catched inside fetcher
-            log('5.1.A. request was not aborted');
-            // Check that new request was not created
-            if (ctx.abortController === abortController) {
-              log('5.1.A.1. Done');
-              delete ctx.abortController;
-              dispatch(create('_done', params, fetcherResult));
-              if (options.onSuccess) {
-                options.onSuccess(dispatch, params, fetcherResult);
-              }
-            }
-          } catch (e) {
-            if (isAbortError(e)) {
-              log('5.1.B. request was aborted');
-              log('5.1.B.1. Error');
-              dispatch(create('_error', params, ABORT_ERROR_MESSAGE));
-            } else if (ctx.abortController === abortController) {
-              delete ctx.abortController;
-              log('5.1.C Not abort error');
-              console.error(`[${name}]:`, e);
-              const errorMessage = isErrorWithMessage(e)
-                ? e.message
-                : typeof e === 'string'
-                  ? e
-                  : 'Unknown';
-              dispatch(create('_error', params, errorMessage));
-            }
-          }
-        });
-      });
-
-      // Force refetch, useful for polling
-      onAction('refetch', () => {
-        schedule((dispatch, ctx: Context) => {
-          if (state.dirty && !state.loading) {
-            dispatch(create('request', newState.lastParams!));
-          } else {
-            console.error(`[${name}]:`, 'Do not call refetch before request');
-          }
-        });
-      });
-
-      onAction('cancel', () => {
-        schedule(async (dispatch, ctx: Context) => {
-          if (ctx.abortController) {
-            ctx.abortController.abort();
-            delete ctx.abortController;
-          }
-        });
-      });
-
-      onAction('_loading', ({ params }) => {
-        newState.loading = true;
-        newState.error = null;
-        newState.lastParams = params;
-      });
-
-      onAction('_error', ({ params, error }) => {
-        newState.error = error;
-        newState.lastParams = params;
-        newState.loading = false;
-      });
-
-      onAction('_done', ({ data, params }) => {
-        newState.data = data;
-        newState.error = null;
-        newState.loading = false;
-        newState.lastParams = params;
-      });
-
-      if (deps.depsAtom) {
-        onChange('depsAtom', (depsAtomState: unknown) => {
-          if (isObject(depsAtomState)) {
-            // Deps is resource atom-like object
-            if (options.inheritState) {
-              newState.loading = depsAtomState.loading || newState.loading;
-              newState.error = depsAtomState.error || newState.error;
-            }
-
-            if (isAtomLike(depsAtomState)) {
-              if (!depsAtomState.loading && !depsAtomState.error && depsAtomState.dirty) {
-                schedule((dispatch) =>
-                  dispatch(create('request', depsAtomState.data as any)),
-                );
-              }
-            } else {
-              // Is plain object
-              schedule((dispatch) => dispatch(create('request', depsAtomState as any)));
-            }
-          } else {
-            // Deps is primitive
-            if (depsAtomState !== null)
-              schedule((dispatch) => dispatch(create('request', depsAtomState as any)));
-            else {
-              console.warn(
-                `Resource atom with name ${name} skips running as its dependency state ${deps?.depsAtom?.id} is null`,
-              );
-            }
-          }
-        });
-      } else {
-        if (options.auto && !newState.dirty) {
-          schedule((dispatch) => dispatch(create('request', null as any)));
-        }
-      }
-
-      return newState;
-    },
+  const asyncAtom = atom<State>(
     {
-      id: getUniqueId(name),
-      store: options.store,
-      decorators: [],
+      loading: false,
+      data: null,
+      error: null,
+      lastParams: null,
+      dirty: false,
     },
+    getUniqueId(name),
   );
 
-  return resourceAtom;
+  let abortController: AbortController | null = null;
+  let deferredCancel: DeferredPromise<void>;
+
+  const requestAction = action(async (ctx, params) => {
+    log('request');
+    // In case we already have request action in progress, cancel it
+    if (abortController) {
+      log('wait cancel');
+      await cancelAction(ctx);
+      log('after cancel');
+    }
+    // Set loading state
+
+    asyncAtom(ctx, (state) =>
+      logStateChange(state, {
+        ...state,
+        error: null,
+        lastParams: params,
+        dirty: true, // for unblock refetch
+        loading: true,
+      }),
+    );
+    log('after async atom state change');
+    try {
+      abortController = new AbortController();
+      const data = await ctx.schedule(() => {
+        if (abortController) {
+          return abortable(abortController, fetcher(params, abortController));
+        } else {
+          throw Error('abortController was reset before it was used');
+        }
+      });
+      log('after fetcher');
+      // Inside the fetcher's logic, abort error could have been caught accidentally
+      // Here I rethrow abort error to be sure it will be processed
+      abortController.signal.throwIfAborted();
+      // Set done state
+      asyncAtom(ctx, (state) =>
+        logStateChange(state, { ...state, data, lastParams: params, loading: false }),
+      );
+    } catch (error) {
+      log('error', error);
+      log('set error state');
+      asyncAtom(ctx, (state) =>
+        logStateChange(state, {
+          ...state,
+          lastParams: params,
+          loading: false,
+          error: generateErrorMessage(error),
+        }),
+      );
+      if (isAbortError(error)) {
+        deferredCancel?.resolve();
+      }
+    } finally {
+      abortController = null;
+    }
+  });
+
+  const cancelAction = action(async (ctx) => {
+    if (abortController) {
+      // If previous request is active
+      deferredCancel = deferred();
+
+      await ctx.schedule(async () => {
+        // Cancel it (that trigger catch section in previous request)
+        abortController?.abort();
+        // Wait for abort state;
+        return await deferredCancel.promise;
+      });
+    }
+  });
+
+  const refetchAction = action((ctx) => {
+    const { lastParams, dirty, loading } = ctx.get(asyncAtom);
+    if (dirty) {
+      !loading && requestAction(ctx, lastParams);
+    } else {
+      console.error(`[${name}]:`, 'Do not call refetch before request');
+    }
+  });
+
+  const actions = {
+    request: requestAction,
+    refetch: refetchAction,
+    cancel: cancelAction,
+  };
+
+  if (depsAtom) {
+    const onChange = (ctx: Ctx, depsAtomState: unknown) => {
+      console.debug('onChange', depsAtomState);
+      if (isObject(depsAtomState)) {
+        // If Deps atom looks like resource
+        console.debug('options.inheritState', options.inheritState);
+        if (options.inheritState) {
+          // Use his properties for state
+          asyncAtom(ctx, (state) => ({
+            ...state,
+            loading: depsAtomState.loading || state.loading,
+            error: depsAtomState.error || state.error,
+          }));
+        }
+        if (isAtomLike(depsAtomState)) {
+          if (!depsAtomState.loading && !depsAtomState.error && depsAtomState.dirty) {
+            requestAction(ctx, depsAtomState.data);
+          }
+        } else {
+          // Is plain object
+          requestAction(ctx, depsAtomState);
+        }
+      } else {
+        // Deps is primitive
+        if (depsAtomState !== null) {
+          requestAction(ctx, depsAtomState);
+        } else {
+          console.warn(
+            `Resource atom with name ${name} skips running as its dependency state ${depsAtom?.id} is null`,
+          );
+        }
+      }
+    };
+
+    // Request data with current deps state
+    onChange(options.store.v3ctx, options.store.getState(depsAtom));
+
+    // Call request action when deps changes
+    depsAtom.v3atom.onChange(onChange);
+    return v3toV2<State, AsyncAtomDeps<D, F>>(asyncAtom, actions, options.store);
+  } else {
+    if (options.auto) {
+      requestAction(options.store.v3ctx, null);
+    }
+  }
+
+  return v3toV2<State, AsyncAtomDeps<D, F>>(asyncAtom, actions, options.store);
 }
